Extracted comparer cache to own class
