+++ FileHelpers/Core/LineInfo.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics;
 using System.Globalization;
+using FileHelpers.Core;
 using FileHelpers.Helpers;
 
 namespace FileHelpers
@@ -202,7 +203,7 @@ namespace FileHelpers
             mCurrentPos = 0;
         }
 
-        private static readonly CompareInfo mCompare = StringHelper.CreateComparer();
+        private static readonly CompareInfo mCompare = ComparerCache.CreateComparer();
 
         /// <summary>
         /// Find the location of the next string in record
